2 files- 1) sending messages from your app 2)receiving messages from phone and replying back to the messages
I have added 2 file for performing two different tasks. the file under outbound messages only folder is able to send message whenever the file is run on the node command prompt. The other file under the inbound and outbound messaging, is able to receive messages from your phone and reply to the messages